Refactored so that this actually compiles
